+++ tests/test_Backend_qlazy_mpstate_simulator.py
@@ -911,6 +911,39 @@ class TestBackend_expect_mps_simulator(unittest.TestCase):
         self.assertEqual(abs(expval_1.real - expval_2.real) < 0.05, True)
 
 #
+# init option
+#
+
+class TestBackend_init_option_qstate_simulator(unittest.TestCase):
+    """ test 'Backend' : init option
+    """
+
+    def test_init_option_run(self):
+        """test 'init'
+        """
+        bk = Backend(product='qlazy', device='mps_simulator')
+        qc_A = QCirc().h(0).h(1).crx(0, 2, phase=0.7).crz(1,2, phase=0.3)
+        qc_B = QCirc().cx(0,1).crx(1,2).crz(2,0, phase=0.9).rx(2, phase=0.4)
+        qc = qc_A + qc_B
+        mps_expect = bk.run(qcirc=qc, out_state=True).mpstate
+        mps_ini = bk.run(qcirc=qc_A, out_state=True).mpstate
+        mps_actual = bk.run(init=mps_ini, qcirc=qc_B, out_state=True).mpstate
+        self.assertEqual(equal_mpstates(mps_expect, mps_actual), True)
+
+    def test_init_option_expect(self):
+        """test 'init'
+        """
+        ob = X(0) * X(1) + Y(1) * Y(2) + Z(0) * Z(2)
+        bk = Backend(product='qlazy', device='mps_simulator')
+        qc_A = QCirc().h(0).h(1).crx(0, 2, phase=0.7).crz(1,2, phase=0.3)
+        qc_B = QCirc().cx(0,1).crx(1,2).crz(2,0, phase=0.9).rx(2, phase=0.4)
+        qc = qc_A + qc_B
+        ev_expect = bk.expect(qcirc=qc, observable=ob, precise=True)
+        mps_ini = bk.run(qcirc=qc_A, out_state=True).mpstate
+        ev_actual = bk.expect(init=mps_ini, qcirc=qc_B, observable=ob, precise=True)
+        self.assertEqual(abs(ev_expect - ev_actual) < EPS, True)
+        
+#
 # inheritance
 #